+++ src/main/java/com/novaventure/survivalessentials/registry/Fluids/RottingFluid/RottingFluid.java
@@ -3,8 +3,6 @@ package com.novaventure.survivalessentials.registry.Fluids.RottingFluid;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.block.BlockState;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.fluid.Fluid;
 import net.minecraft.fluid.FluidState;
 import net.minecraft.item.Item;
@@ -88,7 +86,7 @@ public abstract class RottingFluid extends AbstractModFluid {
     public void randomDisplayTick(BlockState state, World world, BlockPos pos, Random random) {
         if (random.nextInt(12) == 0) {
             double d = (double) pos.getX() + 0.5D;
-            double e = (double) pos.getY();
+            double e = pos.getY();
             double f = (double) pos.getZ() + 0.5D;
             world.playSound((double)pos.getX() + 0.5D, (double)pos.getY() + 0.5D, (double)pos.getZ() + 0.5D, SoundEvents.BLOCK_FIRE_AMBIENT, SoundCategory.BLOCKS, 1.0F + random.nextFloat(), random.nextFloat() * 0.7F + 0.3F, false);
             world.addParticle(ParticleTypes.SMOKE, d, e + 1.1D, f, 0.0D, 0.0D, 0.0D);
